Don't send the request to the flutter api for subframes. Also improve tests and comments
