+++ packages/webview_flutter/webview_flutter_android/android/src/test/java/io/flutter/plugins/webviewflutter/WebViewClientImplTest.java
@@ -4,18 +4,186 @@
 
 package io.flutter.plugins.webviewflutter;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
+import static org.mockito.Mockito.when;
+
+import android.net.Uri;
+import android.webkit.WebResourceRequest;
+import android.webkit.WebResourceResponse;
+import android.webkit.WebView;
 import android.webkit.WebViewClient;
+import androidx.annotation.NonNull;
+import io.flutter.plugins.webviewflutter.WebViewClientHostApiImpl.WebViewClientCreator;
+import java.util.HashMap;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnit;
+import org.mockito.junit.MockitoRule;
 
-public class WebViewClientImplTest extends WebViewClientTest {
+public class WebViewClientImplTest {
+  @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
 
-  @Override
-  WebViewClient createInstance(WebViewClientFlutterApiImpl flutterApi) {
-    return new WebViewClientHostApiImpl.WebViewClientImpl(flutterApi);
+  @Mock public WebViewClientFlutterApiImpl mockFlutterApi;
+
+  @Mock public WebView mockWebView;
+
+  InstanceManager instanceManager;
+  WebViewClientHostApiImpl hostApiImpl;
+  WebViewClientHostApiImpl.WebViewClientImpl webViewClient;
+
+  @Before
+  public void setUp() {
+    instanceManager = InstanceManager.create(identifier -> {});
+
+    final WebViewClientCreator webViewClientCreator =
+        new WebViewClientCreator() {
+          @Override
+          @NonNull
+          public WebViewClient createWebViewClient(
+              @NonNull WebViewClientFlutterApiImpl flutterApi) {
+            webViewClient = new WebViewClientHostApiImpl.WebViewClientImpl(flutterApi);
+            return webViewClient;
+          }
+        };
+
+    hostApiImpl =
+        new WebViewClientHostApiImpl(instanceManager, webViewClientCreator, mockFlutterApi);
+    hostApiImpl.create(1L);
   }
 
-  @Override
-  void setReturnValueForShouldOverrideUrlLoading(WebViewClient client, boolean value) {
-    ((WebViewClientHostApiImpl.WebViewClientImpl) client)
-        .setReturnValueForShouldOverrideUrlLoading(value);
+  @After
+  public void tearDown() {
+    instanceManager.stopFinalizationListener();
+  }
+
+  @Test
+  public void onPageStarted() {
+    webViewClient.onPageStarted(mockWebView, "https://www.google.com", null);
+    verify(mockFlutterApi)
+        .onPageStarted(eq(webViewClient), eq(mockWebView), eq("https://www.google.com"), any());
+  }
+
+  @Test
+  public void onReceivedError() {
+    webViewClient.onReceivedError(mockWebView, 32, "description", "https://www.google.com");
+    verify(mockFlutterApi)
+        .onReceivedError(
+            eq(webViewClient),
+            eq(mockWebView),
+            eq(32L),
+            eq("description"),
+            eq("https://www.google.com"),
+            any());
+  }
+
+  @Test
+  public void urlLoading() {
+    webViewClient.shouldOverrideUrlLoading(mockWebView, "https://www.google.com");
+    verify(mockFlutterApi)
+        .urlLoading(eq(webViewClient), eq(mockWebView), eq("https://www.google.com"), any());
+  }
+
+  @Test
+  public void urlLoadingForMainFrame() {
+    webViewClient.setReturnValueForShouldOverrideUrlLoading(false);
+
+    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
+    when(mockRequest.isForMainFrame()).thenReturn(true);
+
+    assertFalse(webViewClient.shouldOverrideUrlLoading(mockWebView, mockRequest));
+    verify(mockFlutterApi)
+        .requestLoading(eq(webViewClient), eq(mockWebView), eq(mockRequest), any());
+  }
+
+  @Test
+  public void urlLoadingForMainFrameWithOverride() {
+    webViewClient.setReturnValueForShouldOverrideUrlLoading(true);
+
+    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
+    when(mockRequest.isForMainFrame()).thenReturn(true);
+
+    assertTrue(webViewClient.shouldOverrideUrlLoading(mockWebView, mockRequest));
+    verify(mockFlutterApi)
+        .requestLoading(eq(webViewClient), eq(mockWebView), eq(mockRequest), any());
+  }
+
+  @Test
+  public void urlLoadingNotForMainFrame() {
+    webViewClient.setReturnValueForShouldOverrideUrlLoading(false);
+
+    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
+    when(mockRequest.isForMainFrame()).thenReturn(false);
+
+    assertFalse(webViewClient.shouldOverrideUrlLoading(mockWebView, mockRequest));
+    verifyNoInteractions(mockFlutterApi);
+  }
+
+  @Test
+  public void urlLoadingNotForMainFrameWithOverride() {
+    webViewClient.setReturnValueForShouldOverrideUrlLoading(true);
+
+    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
+    when(mockRequest.isForMainFrame()).thenReturn(false);
+
+    assertFalse(webViewClient.shouldOverrideUrlLoading(mockWebView, mockRequest));
+    verifyNoInteractions(mockFlutterApi);
+  }
+
+  @Test
+  public void convertWebResourceRequestWithNullHeaders() {
+    final Uri mockUri = mock(Uri.class);
+    when(mockUri.toString()).thenReturn("");
+
+    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
+    when(mockRequest.getMethod()).thenReturn("method");
+    when(mockRequest.getUrl()).thenReturn(mockUri);
+    when(mockRequest.isForMainFrame()).thenReturn(true);
+    when(mockRequest.getRequestHeaders()).thenReturn(null);
+
+    final GeneratedAndroidWebView.WebResourceRequestData data =
+        WebViewClientFlutterApiImpl.createWebResourceRequestData(mockRequest);
+    assertEquals(data.getRequestHeaders(), new HashMap<String, String>());
+  }
+
+  @Test
+  public void doUpdateVisitedHistory() {
+    webViewClient.doUpdateVisitedHistory(mockWebView, "https://www.google.com", true);
+    verify(mockFlutterApi)
+        .doUpdateVisitedHistory(
+            eq(webViewClient), eq(mockWebView), eq("https://www.google.com"), eq(true), any());
+  }
+
+  @Test
+  public void onReceivedHttpError() {
+    final Uri mockUri = mock(Uri.class);
+    when(mockUri.toString()).thenReturn("");
+
+    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
+    when(mockRequest.getMethod()).thenReturn("method");
+    when(mockRequest.getUrl()).thenReturn(mockUri);
+    when(mockRequest.isForMainFrame()).thenReturn(true);
+    when(mockRequest.getRequestHeaders()).thenReturn(null);
+
+    final WebResourceResponse mockResponse = mock(WebResourceResponse.class);
+    when(mockResponse.getStatusCode()).thenReturn(404);
+
+    webViewClient.onReceivedHttpError(mockWebView, mockRequest, mockResponse);
+    verify(mockFlutterApi)
+        .onReceivedHttpError(
+            eq(webViewClient),
+            eq(mockWebView),
+            any(WebResourceRequest.class),
+            any(WebResourceResponse.class),
+            any());
   }
 }
