+++ packages/webview_flutter/webview_flutter_android/android/src/test/java/io/flutter/plugins/webviewflutter/WebViewClientTest.java
@@ -4,24 +4,13 @@
 
 package io.flutter.plugins.webviewflutter;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
-import android.net.Uri;
-import android.webkit.WebResourceRequest;
-import android.webkit.WebResourceResponse;
-import android.webkit.WebView;
 import android.webkit.WebViewClient;
 import androidx.annotation.NonNull;
 import io.flutter.plugins.webviewflutter.WebViewClientHostApiImpl.WebViewClientCompatImpl;
 import io.flutter.plugins.webviewflutter.WebViewClientHostApiImpl.WebViewClientCreator;
-import java.util.HashMap;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
@@ -30,22 +19,14 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-public abstract class WebViewClientTest {
+public class WebViewClientTest {
   @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
 
   @Mock public WebViewClientFlutterApiImpl mockFlutterApi;
 
-  @Mock public WebView mockWebView;
-
-  @Mock public WebViewClientCompatImpl mockWebViewClient;
-
   InstanceManager instanceManager;
   WebViewClientHostApiImpl hostApiImpl;
-  WebViewClient webViewClient;
-
-  abstract WebViewClient createInstance(WebViewClientFlutterApiImpl flutterApi);
-
-  abstract void setReturnValueForShouldOverrideUrlLoading(WebViewClient client, boolean value);
+  WebViewClientCompatImpl webViewClient;
 
   @Before
   public void setUp() {
@@ -57,7 +38,7 @@ public abstract class WebViewClientTest {
           @NonNull
           public WebViewClient createWebViewClient(
               @NonNull WebViewClientFlutterApiImpl flutterApi) {
-            webViewClient = createInstance(flutterApi);
+            webViewClient = (WebViewClientCompatImpl) super.createWebViewClient(flutterApi);
             return webViewClient;
           }
         };
@@ -73,100 +54,8 @@ public abstract class WebViewClientTest {
   }
 
   @Test
-  public void onPageStarted() {
-    webViewClient.onPageStarted(mockWebView, "https://www.google.com", null);
-    verify(mockFlutterApi)
-        .onPageStarted(eq(webViewClient), eq(mockWebView), eq("https://www.google.com"), any());
-  }
-
-  @Test
-  @SuppressWarnings("deprecation")
-  public void onReceivedError() {
-    webViewClient.onReceivedError(mockWebView, 32, "description", "https://www.google.com");
-    verify(mockFlutterApi)
-        .onReceivedError(
-            eq(webViewClient),
-            eq(mockWebView),
-            eq(32L),
-            eq("description"),
-            eq("https://www.google.com"),
-            any());
-  }
-
-  @Test
-  @SuppressWarnings("deprecation")
-  public void urlLoading() {
-    webViewClient.shouldOverrideUrlLoading(mockWebView, "https://www.google.com");
-    verify(mockFlutterApi)
-        .urlLoading(eq(webViewClient), eq(mockWebView), eq("https://www.google.com"), any());
-  }
-
-  @Test
-  public void urlLoadingForMainFrame() {
-    setReturnValueForShouldOverrideUrlLoading(webViewClient, false);
-
-    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
-    when(mockRequest.isForMainFrame()).thenReturn(true);
-
-    assertFalse(webViewClient.shouldOverrideUrlLoading(mockWebView, mockRequest));
-    verify(mockFlutterApi)
-        .requestLoading(eq(webViewClient), eq(mockWebView), eq(mockRequest), any());
-  }
-
-  @Test
-  public void urlLoadingForMainFrameWithOverride() {
-    setReturnValueForShouldOverrideUrlLoading(webViewClient, true);
-
-    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
-    when(mockRequest.isForMainFrame()).thenReturn(true);
-
-    assertTrue(webViewClient.shouldOverrideUrlLoading(mockWebView, mockRequest));
-    verify(mockFlutterApi)
-        .requestLoading(eq(webViewClient), eq(mockWebView), eq(mockRequest), any());
-  }
-
-  @Test
-  public void urlLoadingNotForMainFrame() {
-    setReturnValueForShouldOverrideUrlLoading(webViewClient, false);
-
-    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
-    when(mockRequest.isForMainFrame()).thenReturn(false);
-
-    assertFalse(webViewClient.shouldOverrideUrlLoading(mockWebView, mockRequest));
-    verify(mockFlutterApi)
-        .requestLoading(eq(webViewClient), eq(mockWebView), eq(mockRequest), any());
-  }
-
-  @Test
-  public void urlLoadingNotForMainFrameWithOverride() {
-    setReturnValueForShouldOverrideUrlLoading(webViewClient, true);
-
-    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
-    when(mockRequest.isForMainFrame()).thenReturn(false);
-
-    assertFalse(webViewClient.shouldOverrideUrlLoading(mockWebView, mockRequest));
-    verify(mockFlutterApi)
-        .requestLoading(eq(webViewClient), eq(mockWebView), eq(mockRequest), any());
-  }
-
-  @Test
-  public void convertWebResourceRequestWithNullHeaders() {
-    final Uri mockUri = mock(Uri.class);
-    when(mockUri.toString()).thenReturn("");
-
-    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
-    when(mockRequest.getMethod()).thenReturn("method");
-    when(mockRequest.getUrl()).thenReturn(mockUri);
-    when(mockRequest.isForMainFrame()).thenReturn(true);
-    when(mockRequest.getRequestHeaders()).thenReturn(null);
-
-    final GeneratedAndroidWebView.WebResourceRequestData data =
-        WebViewClientFlutterApiImpl.createWebResourceRequestData(mockRequest);
-    assertEquals(data.getRequestHeaders(), new HashMap<String, String>());
-  }
-
-  @Test
   public void setReturnValueForShouldOverrideUrlLoading() {
+    final WebViewClientCompatImpl mockWebViewClient = mock();
     final WebViewClientHostApiImpl webViewClientHostApi =
         new WebViewClientHostApiImpl(
             instanceManager,
@@ -184,37 +73,5 @@ public abstract class WebViewClientTest {
     webViewClientHostApi.setSynchronousReturnValueForShouldOverrideUrlLoading(2L, false);
 
     verify(mockWebViewClient).setReturnValueForShouldOverrideUrlLoading(false);
-  }
-
-  @Test
-  public void doUpdateVisitedHistory() {
-    webViewClient.doUpdateVisitedHistory(mockWebView, "https://www.google.com", true);
-    verify(mockFlutterApi)
-        .doUpdateVisitedHistory(
-            eq(webViewClient), eq(mockWebView), eq("https://www.google.com"), eq(true), any());
-  }
-
-  @Test
-  public void onReceivedHttpError() {
-    final Uri mockUri = mock(Uri.class);
-    when(mockUri.toString()).thenReturn("");
-
-    final WebResourceRequest mockRequest = mock(WebResourceRequest.class);
-    when(mockRequest.getMethod()).thenReturn("method");
-    when(mockRequest.getUrl()).thenReturn(mockUri);
-    when(mockRequest.isForMainFrame()).thenReturn(true);
-    when(mockRequest.getRequestHeaders()).thenReturn(null);
-
-    final WebResourceResponse mockResponse = mock(WebResourceResponse.class);
-    when(mockResponse.getStatusCode()).thenReturn(404);
-
-    webViewClient.onReceivedHttpError(mockWebView, mockRequest, mockResponse);
-    verify(mockFlutterApi)
-        .onReceivedHttpError(
-            eq(webViewClient),
-            eq(mockWebView),
-            any(WebResourceRequest.class),
-            any(WebResourceResponse.class),
-            any());
   }
 }
